+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.1.2
+- Add json params to departments, roles and apps
+- Add Auth and User
+- Support NormalAuth and NormalUser
+
 ## 0.1.1
 - Add params to users